+++ lua/obsidian/lsp/handlers/document_symbol.lua
@@ -1,38 +1,27 @@
 local obsidian = require "obsidian"
+local Range = require "obsidian.range"
 
 ---@param _ lsp.DocumentSymbolParams
 return function(_, handler)
   ---@type lsp.DocumentSymbol[]
   local symbols = {}
-  local lookup = {}
 
-  local note = obsidian.Note.from_buffer(0, { collect_anchor_links = true })
+  local note = obsidian.Note.from_buffer(0, { collect_sections = true })
   if not note then -- HACK: somehow DocumentSymbol gets called with no valid note
     return handler(nil, symbols)
   end
 
-  for _, anchor in pairs(note.anchor_links) do
-    local display = string.rep("#", anchor.level) .. " " .. anchor.header
-    local rge = {
-      start = { line = anchor.line - 1, character = 0 },
-      ["end"] = { line = anchor.line - 1, character = 0 },
-    }
-    if not lookup[anchor.line] then
-      local symbol = {
-        name = display,
+  for _, section in ipairs(note.sections or {}) do
+    if section.header then
+      symbols[#symbols + 1] = {
+        name = string.rep("#", section.level) .. " " .. section.header,
         kind = 1,
         filename = note.path.filename,
-        range = rge,
-        selectionRange = rge,
+        range = Range.to_lsp(section.range),
+        selectionRange = Range.to_lsp(section.heading_range),
       }
-      symbols[#symbols + 1] = symbol
-      lookup[anchor.line] = true
     end
   end
-
-  table.sort(symbols, function(a, b)
-    return a.range.start.line < b.range.start.line
-  end)
 
   handler(nil, symbols)
 end